Refactor: drop the ``versioned_`` prefix + split persistence mixins
The ``versioned_*`` prefix was useful during Phase 1 to coexist with
the old non-versioned store modules.  Those are gone now, so the
prefix is just noise — every spec store IS the versioned spec store.

Module layout after this refactor:

  base_store.py   — ``SpecStore`` ABC + ``_StoreState`` + the
                    ``migrate_legacy_mapping`` helper.  Knows *nothing*
                    about file or redis persistence.
  file_store.py   — ``FileSpecStore`` mixin (tmp+rename on every save).
  redis_store.py  — ``RedisSpecStore`` mixin (snapshot read/write).
  store.py        — ``AgentStore`` (spec-specific logic) + concrete
                    ``FileAgentStore`` / ``RedisAgentStore`` composed
                    from the persistence mixins.
  team_store.py   — ``TeamStore`` + ``FileTeamStore`` / ``RedisTeamStore``
                    following the same pattern.

Adding a new persistence backend (Postgres, SQLite, …) is now three
lines: write a mixin with ``_load_state`` / ``_save_state`` /
``migrate_from_legacy``, subclass the agent/team store with it, register
it in ``AGENT_STORE_ALIASES``.  No edits to the base ABC.

Also strips the redundant ``import X as X`` aliases that came in
during the Phase 1→2 handoff (they were only needed while both names
coexisted).

Tests pass unchanged; the two ``test_versioned_*`` files were renamed
to drop the prefix.  No behavior change — this is pure cleanup.
